Refactor: put functions into other modules
