+++ src/common/evaluation/QA/eval-depthmap-height/src/constants.py
@@ -1,3 +1,4 @@
 from pathlib import Path
 
 REPO_DIR = Path(__file__).parents[6].absolute()
+DATA_DIR_ONLINE_RUN = Path("/tmp/data/")
