+++ src/common/evaluation/QA/eval-depthmap-height/src/utils.py
@@ -1,10 +1,26 @@
+import datetime
 import os
+from pathlib import Path
 import pickle
 
-from azureml.core import Experiment, Run
+from azureml.core import Experiment, Run, Workspace
 import glob2 as glob
 import numpy as np
 import pandas as pd
+
+
+def download_dataset(workspace: Workspace, dataset_name: str, dataset_path: str):
+    print("Accessing dataset...")
+    if os.path.exists(dataset_path):
+        return
+    dataset = workspace.datasets[dataset_name]
+    print("Downloading dataset.. Current date and time: ", datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
+    dataset.download(target_path=dataset_path, overwrite=False)
+    print("Finished downloading, Current date and time: ", datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
+
+
+def get_dataset_path(data_dir: Path, dataset_name: str):
+    return str(data_dir / dataset_name)
 
 
 def preprocess_depthmap(depthmap):
